agregacion de dominio al usuario

+++ src/main/java/edu/student/itson/bda/ticketwizard/entidades/Usuario.java
@@ -81,7 +81,7 @@ public class Usuario {
         this.email = email;
     }
 
-    public float getSaldo() {
+    public int getSaldo() {
         return saldo;
     }
